pyproject.toml
[tool.poetry]
name = "deneb"
version = "0.1.0"
authors = ["anonymous <anonymous@example.com>"]
description = "ECCV"
license = "BSD-3-Clause-Clear"
readme = "README.md"
classifiers = [
    'Development Status :: 4 - Beta',
    'Environment :: Console',
    'Intended Audience :: Science/Research',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
]
packages = [
    {include = "deneb"},
]
include = [
    "LICENSE",
    "pyproject.toml",
    "CONTRIBUTING.md"
]

[tool.poetry.scripts]
deneb = 'deneb.cli:deneb'

[tool.poetry.dependencies]
python = ">3.9,<=3.10.0"
PyYAML = "5.3.*"
#numpy = "<=1.19"
protobuf="<3.20.0"
pandas = "^1.0.0"
#fairseq = "0.9.0"
transformers = "4.*"
pytorch-lightning = "<=1.3"
pytorch-nlp = "0.5.0"
fsspec = "0.8.7"
psycopg2-binary = "^2.9.6"
torch = {version = "^2.1.0+cu118", source = "torch_cu118"}
torchvision = {version = "^0.16.0+cu118", source = "torch_cu118"}
click = "^8.1.7"
clip = { path = "./CLIP" }
ftfy = "^6.1.1"
regex = "^2023.10.3"
wandb = "^0.15.12"
scikit-learn = "^1.3.2"
fairseq = "^0.12.2"
matplotlib = "^3.8.1"
pycocotools = "^2.0.7"
pycocoevalcap = "^1.2"
termcolor = "^2.3.0"
open-clip-torch = "^2.24.0"
smart-open = "^7.0.1"

[tool.poetry.dev-dependencies]
sphinx-markdown-tables = "0.0.15"
coverage = "^5.5"

[[tool.poetry.source]]
name = "torch_cu118"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
